+++ src/ui/main_window.py
@@ -46,6 +46,7 @@ class MainWindow(QMainWindow):
         self.operation_buttons['harvest_metadata'] = QPushButton("Harvest Metadata")
         self.operation_buttons['created_to_mod'] = QPushButton("Created to Mod")
         self.operation_buttons['auto_rename'] = QPushButton("Auto Rename")
+        self.operation_buttons['separate_media'] = QPushButton("Separate Media")
         # self.operation_buttons['auto_select'] = QPushButton("Auto Select")
         # self.operation_buttons['import'] = QPushButton("Import")
         # self.operation_buttons['rotate'] = QPushButton("Rotate")
@@ -57,7 +58,6 @@ class MainWindow(QMainWindow):
         # self.operation_buttons['archive'] = QPushButton("Archive")
         # self.operation_buttons['cleanup'] = QPushButton("Cleanup")
         # self.operation_buttons['make_sequential'] = QPushButton("Make Sequential")
-        # self.operation_buttons['separate_video'] = QPushButton("Separate Video")
 
         self.group_operations = QGroupBox("Operations")
         layout_buttons = QVBoxLayout(self.group_operations)